testpure1

+++ src/utils.c
@@ -60,7 +60,7 @@ void *__capability codecap_create(void *sandbox_base, void *sandbox_end, int dis
 	if(disable_lwec)
 		extra_perm = 0;
 
-	codecap = cheri_codeptrperm(sandbox_base, ((size_t) sandbox_end - (size_t) sandbox_base),
+	codecap = cheri_codeptrperm((unsigned long)sandbox_base, ((size_t) sandbox_end - (size_t) sandbox_base),
 				    CHERI_PERM_GLOBAL | CHERI_PERM_LOAD | CHERI_PERM_STORE | CHERI_PERM_EXECUTE | CHERI_PERM_CCALL | CHERI_PERMS_HWALL | extra_perm);
 
 	return (codecap);
@@ -74,7 +74,7 @@ void *__capability pure_codecap_create(void *sandbox_base, void *sandbox_end, in
 	if(disable_lwec)
 		extra_perm = 0;
 
-	codecap = cheri_codeptrperm(sandbox_base, ((size_t) sandbox_end - (size_t) sandbox_base),
+	codecap = cheri_codeptrperm((unsigned long)sandbox_base, ((size_t) sandbox_end - (size_t) sandbox_base),
 				    CHERI_PERM_GLOBAL | CHERI_PERM_LOAD | CHERI_PERM_LOAD_CAP | CHERI_PERM_STORE | CHERI_PERM_EXECUTE | CHERI_PERM_CCALL | CHERI_PERMS_HWALL | extra_perm);
 
 	codecap = cheri_capmode(codecap);
